requirements.txt
PyPDF2==3.0.1
reportlab==4.0.4
Pillow==10.0.1
streamlit==1.28.1
pdf2image==1.16.3
PyMuPDF==1.23.8
numpy==1.24.3 